requirements.txt
fastapi==0.121.0
httpx==0.28.1
pydantic==2.12.4
pymongo==4.15.3
python-dotenv==1.2.1
Faker==33.1.0
sentence-transformers>=2.2.0
anthropic>=0.18.0
numpy>=1.24.0
